# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Boulder Buddy API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3500",
    "basePath": "/api/v1",
    "paths": {
...
        "/mesos": {
            "get": {
                "description": "get all mesocyles for current user",
                "tags": [
                    "Mesos"
                ],
                "summary": "Retrieve mesocyles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.Mesocycle"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new mesocycle",
                "tags": [
                    "Mesos"
                ],
                "summary": "Create a mesocyle",
                "parameters": [
                    {
                        "description": "new mesocycle info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.Mesocycle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Mesocycle"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.errorCode"
                        }
                    ]
                },
                "fields": {
                    "description": "Some context of error",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Human-readable message.",
                    "type": "string"
                },
                "op": {
                    "description": "Logical operation.",
                    "type": "string"
                },
                "stack": {
                    "description": "Wrapped error"
                }
            }
        },
...
        "handlers.Mesocycle": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "integer"
                },
                "goal": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "numMicros": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "integer"
                }
            }
        },
...
        "utils.errorCode": {
            "type": "string",
            "enum": [
                "conflict",
                "internal",
                "invalid",
                "not_found",
                "not_modified",
                "already_exists",
                "permission_denied",
                "unauthenticated",
                "cannot_decode",
                "cannot_encode",
                "cannot_parse",
                "undefined_behavior",
                "unsupported",
                "test_error_code"
            ],
            "x-enum-varnames": [
                "ECONFLICT",
                "EINTERNAL",
                "EINVALID",
                "ENOTFOUND",
                "ENOTMODIFIED",
                "EALREADYEXISTS",
                "EPERMISSIONDENIED",
                "EUNAUTHENTICATED",
                "ECANNOTDECODE",
                "ECANNOTENCODE",
                "ECANNOTPARSE",
                "EBEHAVIOUR",
                "EUNSUPPORTED",
                "ETEST"
            ]
        }
    }
}